Scripts06-Analytics/Program/requirements.txt
snowflake-connector-python[pandas]
psycopg2-binary
tqdm
python-dotenv

